+++ agent/listener/certificate.py
@@ -36,9 +36,7 @@ def create_self_signed_cert(cert_dir, cert_file, key_file):
         cert.set_pubkey(k)
         cert.sign(k, 'sha256')
 
-        open(target_cert, "wt").write(
-            OpenSSL.crypto.dump_certificate(OpenSSL.crypto.FILETYPE_PEM, cert))
-        open(target_key, "wt").write(
-            OpenSSL.crypto.dump_privatekey(OpenSSL.crypto.FILETYPE_PEM, k))
+        open(target_cert, "w").write(OpenSSL.crypto.dump_certificate(OpenSSL.crypto.FILETYPE_PEM, cert))
+        open(target_key, "w").write(OpenSSL.crypto.dump_privatekey(OpenSSL.crypto.FILETYPE_PEM, k))
 
     return target_cert, target_key